requirements.txt
# Basic libraries
numpy==1.24.4
pandas==2.1.2
torch==2.0.1
torchvision==0.15.1
Pillow==10.1.0

# Additional libraries
scikit-learn==1.3.1
matplotlib==3.8.1